Fix table to have fixed lengths
Fixed table to have fixed lengths in percentages relative to the width of page. Fixed the bug where the tr did not match the width of the page. Now, the length of rows in the table are the same as the length of the page, and each column has fixed length.